Refactor service into modular routers
